commons::logging::utils

Function is_directory_writable

pub async fn is_directory_writable(path: &Path) -> LoggingResult<bool>

Available on crate features logging-tokio and logging only.

Checks if a directory is writable.

Arguments

path - A reference to a Path that holds the directory path to check.

Returns

A LoggingResult<bool> which is Ok(true) if the directory is writable, Ok(false) otherwise, or an error if the operation fails.

Errors

This function returns an error if the temporary file used for testing writability cannot be removed.